[Touchable] Expose the press-retention offset as a prop
The press-retention offset defines a rect around a touchable component in which the press is retained. Easiest way to see this is to touch a button in a real navigation bar and slide your finger out of the range and back in. This diff exposes the offset as a prop (I thought pressRetentionOffset was a more informative name than pressRectOffset).