[React Native] #WIP Modify RCTShadowText measure function to honor maxNumberOfLines property
Summary: The `-[NSAttributedString(NSExtendedStringDrawing) boundingRectWithSize:options:context:]` method does not take a maximum number of lines argument and that's why @fkg noticed that the sizing was wrong but the display was right. In this diff, I use the iOS 7+ TextKit API to correctly size the text. In the example Playground.js file (P19784051), the original text sizing implementation took ~50 ms for each of about 10k test iterations and the new implementation (when `maxNumberOfLines` is set) takes ~70 ms. So `maxNumberOfLines` costs ~20 ms of performance. I'm looking into ways to make this better but in the meantime this is what I have.

Test Plan: Run P19784051 after patching this diff.