[ReactNative] Fix crash on image download
Summary:
When opening the paged list view example on UIExplorer it would crash after a
while with a SIGABRT when running callback blocks, copying the blocks array
seems to fix it

Test Plan:
*I was just able to reproduce it on the iPhone 4S*

On UIExplorer open `<ListView> - Paging`, go back to the menu, it'll take a while
to process, open the ListView again. It usually crashes right on the second
opening.