Refactored RCTLog and added facility to prepend extra data to the log message
Summary:
This diff refactors RCTLog to make it easier to use, more flexible and more powerful. I've made the following changes:

1) Logging is now performed on a background thread, which should reduce the overhead
2) Custom logging functions now have access to more data and are more flexible
3) It's now possible to add additional log functions instead of just replacing
4) It's now possible to selectively prepend additional text to log messages

The significance of 4) is that errors logged when setting view properties now have much more context to help debug them: {F22104069}

Discussion:

The way that we forward log messages to JS seems overly complex and expensive. Two questions:

* Why do we do this at all? Where are those messages surfaced? Is it only applicable when running a debug context? If so, should we disable it in the JSCore executor?

* Is there really any benefit to sending an array of log values as individual JSON objects instead of a formatted string? The formatted string would already include the contents of arrays/dictionaries and values of numbers and strings anyway. I propose we do away with this, and just send the formatted string, which would make logging much cheaper.

Also, the logging of invariants from JS back to Obj-C seems to be broken. It's sending a string consisting of comma delimited arguments instead of the pre-formatted error message. Should I file a task to fix this? (the problem seems to be at the JS end)

Test Plan: Run through the example apps and test plan