[AdsManager] Correct back button functionality
Summary: This makes routing contextual, so navigators can communicate with each-other. Navigation requests are passed up the component heirarchy, and the handling flows downward so that the correct navigator performs the operation. This also introduces a navigation interceptor, which allows a component to override navigation behavior in that context. See how it can be used in `AdsManagerTabsModalView.android`

Test Plan:
See video for android back-button testing: https://www.intern.facebook.com/pxlcld/mdKV

Tested nav on Adsman iOS for regressions