[ReactNative] Remove bridge retaining cycles
Summary:
Right now one bridge can be used by N RootViews, If only the views have strong
references to bridge we allow it to be automatically deallocated
once it's not being used by any view anymore (And if not's being retained/
managed by any other object)

Test Plan:
Create a new Bridge on the AppDelegate and don't assign it to any strong
variable, the bridge should be deallocated. Same thing if creating a RootView
that automatically creates a Bridge.

Tested Catalyst and the open source apps to ensure the bridge is not being
deallocated unexpectedly.