Also call the original `console` methods if they exist
Ex. When `console.log` or similar is called, this will call
the original method, which can be quite useful.

For example, under iOS, this will log to the Safari console debugger,
which has an expandable UI for inspecting objects, etc., and is also
just useful if you are using that as a REPL.

I don't believe this incurs a meaningful performance penalty unless
the console is open, but it would be easy to stick behind a flag
if that is a problem.

A slightly different version of this patch was submitted a while ago
and merged but something got screwed up with the merge and this change
was not actually pulled in. (See the weird unrelated commits and changes
attached to the PR #205)