[MAdMan] Dimensions.get('window') considered harmful
Summary:
tl;dr: `Dimensions.get('window')` can change (e.g. on iOS when you receive a phone call) and is not guaranteed to return the available dimensions of your app. In that way, I would consider a misnomer. We tend to use it to mean "screen dimensions", so the `'window'` bit is a bit of a red herring. On Android, screen vs window size is particularly evident as the native buttons are part of the screen on modern phones, but the app can't draw there.

Fortunately, we can usually work around this by using either flex or absolute position and specifying all four corners (or just two, depending on the number of dimensions we need to have filled entirely).

Test Plan: tested all touched places and to the best of my eyesight they did not change