[ReactNative] Avoid navigator scene renders during transition
Summary:
Rendering while transitioning causes ugly flashes and rarely needs to recurse to the scene. This makes it so that renders can happen above the navigator without re-rendering the scenes.

This usually happens when you want to re-render the navigation bar during the transition, like in the case of the AdsManager wizard.

Test Plan: Tested navigator example on iOS. Tested AdsManager on iOS and Android.