[ReactNative] Make RCTRootView propage motions
Summary:
`RCTRootView` was handling the shake motions to show the dev menu, but it wasn't
propagating the cases it didn't handle.
That fixes the problem with MAdMan's bug report screen that wasn't appearing.

Test Plan:
Run MAdMan and hit cmd+ctrl+Z and make sure that the bug report screen appears as expected,
also run UIExplorer and check if the dev menu still works.