Added docs for UITableView vs. <ListView>
When I started out with React Native, I instinctively looked around for a `UITableView` component so I could display stuff in a list. But now I understand that React will never get `UITableView` or `UICollectionView` for reasons mentioned by @vjeux here as a comment on a particular issue: https://github.com/facebook/react-native/issues/499#issuecomment-87821298

I feel that it accurately explains the reason, and should be surfaced up to the docs. Hence adding it here.