[RCTBridge] Clean up reload notification observer
After the RCTBatchedBridge refactor, the `-[NSNotificationCenter addObserver]` and `[removeObserver]` calls got divided between RCTBridge and RCTBatchedBridge. This diff moves `removeObserver` out of RTCBatchedBridge and into `-[RCTBridge dealloc]`. It also adds a check for `isValid` in `-[RCTBridge reload]` so that telling an invalid bridge to reload won't reload it a second time if it becomes valid before the dispatch_async block runs.