Auto-register App binary as module provider
This adds another registration function
`RCTRegisterMainAppModuleProvider(void)` that attempts to register the
binary containing the `main()` function as a React Native module
provider.

This allows a project to link against a static library that contains a
React Native module, without itself providing any modules, or having to
explicitly register as a module provider.  Without this, the static
library containing the module is registered, but the `RCTExport`, etc.
strings are actually linked into the app binary and aren't found.