[RFC][ScrollView] Default scrollsToTop to false
On iOS if you have two scroll views that have `scrollsToTop` set to true then neither scrolls to the top. We can't change that, but I was thinking it might be better to have scroll views opt into the scroll-to-top behavior so that adding new scroll views to your app doesn't disable scroll-to-top.

The motivation for this came from when I embedded a community-authored component that contained a scroll view and noticed that my screen no longer scrolled to the top anymore. The author of the component should have set `scrollsToTop` to false but every other person writing a component with a scroll view is going to have the same problem.

This would be a breaking change since scroll views for scene roots would have to set `scrollsToTop={true}`.