Add a setContextName in JavaScriptCore JSContext's to allow setting of context names
This address https://github.com/facebook/react-native/issues/1083
by letting you set any context's name from JavaScript.
This makes certain things easier to debug.

The reason this is implemented as a native hook instead of going
over the bridge is that it only applies to JSC JSContexts and the
context is needed in that case.

Since this won't necessarily be set in other Executors, you should
only call this by checking to see if it defined first.