[Errors] Include unique IDs in the exception manager methods
The exception manager sometimes passes info about the same error more than once to native--this is because it asynchronously fetches and applies the source map, so it later can provide a more accurate stack trace.

This diff includes IDs (just an incrementing counter) so that the native code--specifically the delegate methods--can associate the error info without source maps with the info w/source maps for the same error. This is useful for error logs.

This is a breaking change because of the new exceptionID argument but I expect it will mostly affect just FB and advanced users who will get a clear compiler error.