[npm] Add the semver caret to package.json deps
The caret is standard for `npm --save` and makes it easier to get bugfixes from package maintainers.

As I understand it, Facebook pins the versions since the packages are manually downloaded. One issue with this is that the dependencies of dependencies aren't pinned so what probably makes the most sense for Facebook is to use `npm shrinkwrap` which recursively pins the versions of all dependencies. How this works is:

1. Someone at Facebook manually downloads the latest version of an npm package they want to use
2. They add the entry to package.json using the caret, ex: `"babel": "^5.6.5"`
3. They run `npm shrinkwrap` to generate npm-shrinkwrap.json so the exact versions of the packages are recorded and if you were to run `npm install` then it would know to get those specific versions (example npm-shrinkwrap.json: https://gist.github.com/ide/a2f6e6818802e39978d5)
4. package.json is synced out to GitHub but npm-shrinkwrap.json is not
5. External users run `npm install` and get the latest compatible versions of all packages

Test Plan: Ran UIExplorer and tested the Chrome debugger.