[Crashfix] Ensure that the image response is non-nil before caching it
If you try to create a cached response from a nil response the app will crash. Looking at the code that uses NSURLSession and NSURLCache, this fix looks correct to me.

Test Plan: Load app with remote images and reload it (which was triggering a craash for me) -- no longer crashes upon reload.