[ListView] Operate on the true scroll responder instead of the scroll component
When composing scroll views, `this.refs[SCROLLVIEW_REF]` may refer to another higher-order scroll component instead of a ScrollView. This can cause issues if you expect to need it to be a ScrollView backed by an RCTScrollView.

The solution is to call `getScrollResponder()` - as long as all higher-order scroll components implement this method, it will make its way down to the true ScrollView, which is what ListView wants here.

Test Plan: Load an app with composed scroll views and no longer get a crash in `calculateChildFrames:(NSNumber *)reactTag callback:(RCTResponseSenderBlock)callback`.