Update RCTLocationObserver.m
New patch set based on comment feedback. 
Removed the _not not_ from code and comments
Added an objectForKey check to avoid crash in case location object is empty
Replaced kCLLocationAccuracyBest with kCLLocationAccuracyBestForNavigation (debate this please. This will eat more battery because it' more CPU intensive, but it is the most accurate navigation option. It's really unfortunate that the web API only provides a boolean here).