[Rendering] Set allowsEdgeAntialiasing for transformed views
By default, the edges of rotated layers on iOS have jagged edges because they are not antialiased. Setting `allowsEdgeAntialiasing` makes them look a lot nicer by smoothing out the jaggies. This is particularly important for UIs like Tinder cards, for example.

Test Plan: Get the transform example working in the UIExplorer and do a manual before & after comparison and see that the jaggies are gone.