[Navigator] Vertically hide disabled scenes and use pointerEvents="none"
Hides disabled scenes using `top` instead of `left`, which fixes a bug with the native UITabBar. When the UITabBar's width is zeroed because the scene has `left: SCREEN_WIDTH, right: 0` applied, this triggers a bug with the kerning of the tab titles. Instead, zeroing the height by setting `top: SCREEN_HEIGHT` avoids the bug.

Also applies `pointerEvents="none"` to disabled scenes so that views in the off-screen scenes definitely don't receive touches, which was occurring before.

Test Plan: Display a UITabBar and push a scene entirely over it. Pop the scene and see that the the tab titles are not scrunched.

Thoroughly test the Navigator example in the UIExplorer and see that scenes properly receive touch events when they are visible.