[Navigator.NavigationBar] Title not being centered properly
This is to fix the NavigationBar title. It is not being is not being
centered properly in Landscape mode.

This is because of commit dbe8e31c209dd1c4a8f48c5e26cbb67edb30d2a7.

Changes were made to remove unnecessary use of absolute screen width,
but looks like title was left out of the equation.

https://github.com/facebook/react-native/commit/dbe8e31c209dd1c4a8f48c5e
26cbb67edb30d2a7