[ScrollView] Add scrollWithoutAnimationTo to Android
`ScrollView.scrollWithoutAnimationTo` is supported on iOS but not Android. This diff adds Android support.

Test Plan: Create a small demo app that has a scroll view that has sufficiently large content so as to be scrollable. Add a button that calls `scrollWithoutAnimationTo(100)` when tapped, and verify that the scroll view immediately jumps to the desired position.