[Tests] Increase yeoman generator timeout
The tests were failing on my MBP till I increased the timeout... this should help Travis without making the tests too slow in case there is a real issue. The timeout is now 5 seconds. If the generator completes before then, the test will proceed. Otherwise the tests will just move on and probably fail (which is what's currently happening).