Temporarily depend on 0.12 Android binaries from JCenter repo
Published version 0.12.0 of the Android binaries. Naming the
version 0.12.0-SNAPSHOT or 0.12.0-rc is not supported,
will come up with a better version scheme for next release.

The version has been published to Maven Central:
https://repo1.maven.org/maven2/com/facebook/react/react-native/

However, after more than an hour it's still not available and
Gradle won't fetch it because it hasn't been mirrored to JCenter:
http://search.maven.org/#search%7Cga%7C1%7Ca%3A%22react-native%22
https://bintray.com/bintray/jcenter/com.facebook.react%3Areact-native/view

As a workaround to make it available now, I published it under my
account which makes it available immediately:
https://bintray.com/mkonicek/maven/react-native/view

I also submitted a request to Bintray to publish the binaries here:
https://bintray.com/bintray/jcenter/com.facebook.react%3Areact-native/view

Once 0.12.0 propagates, we can revert this commit.

None of this affects the npm versioning. The npm version is 0.12.0-rc.