Fixed to use ; instead of : as the path delimiter on Windows, for NDK_MODULE_PATH
Summary: Also fixed the extra slash before first-party in that path (it was ...//first-party
and is now .../first-party).  The $dir funct always produces a result that ends with
a /, so APP_MK_DIR always ends in a / and no extra / is needed when appending to it.
With these two changes all of React Native for Android, including C++, seems to
build OK on Windows :smile:
Closes https://github.com/facebook/react-native/pull/2961

Reviewed By: @​svcscm

Differential Revision: D2512239

Pulled By: @kmagiera

fb-gh-sync-id: 4fbc9fc85bab455b8ed1e07a8ac299a48921753e