[Env] Set up the RN JS environment before requiring other modules
Set up the polyfills for console and process before initializing other modules. Some modules do work (e.g. call `invariant`) when they are first required, so requiring the DefaultInjection module at the beginning ensures we have a consistent JS environment.

Test Plan: There is a packager bug that bundles a copy of `invariant` from npm under the name "invariant", which conflicts with RN's copy of `invariant`. That bug should be fixed, but it revealed another bug in which the JS environment doesn't get set up until a couple modules have already been required.

With this diff, the npm copy of `invariant`, which relies on process.env`, no longer fatals.