Flexible textDecorationLine parsing
Makes textDecorationLine a more flexible value; supports a single string enum value of 'underline', or 'line-through', or, a string-separated set of the same enums, or, an array of the same enum values.

In iOS, this used the existing RCT_MULTI_ENUM_CONVERTER macro, but modified to support the space-separated string values.  In Android, ReactAndroid doesn't support an @ReactProp with multiple potential values, so the conversion is done in a JS-side conversion into an integer.