Remove react-native executable warning
Given that React Native is distributed as two npm packages,
`react-native-cli` and `react-native`, I get the desire to point people
in the right direction of `npm install react-native-cli`’ing when they
want a `react-native` executable – this was [first added in this
commit][c1] and then [restored in this commit][c2].

But if a project installs `react-native` as an npm dependency and
`react-native-cli` as an npm _dev_ dependency, it's the `react-native`
bin that gets executed. For example, with these two dependencies in the
same package.json, the following npm script will source `react-native`
and _not_ `react-native-cli`:

```json
{
  "scripts": {
    "foo": "react-native bundle"
  }
}
```

Not only will the error message be a bit of a red herring, but sourcing
the proper `react-native-cli` bin wil require reaching into
node_modules:

```json
{
  "scripts": {
    "foo": "node_modules/react-native-cli/index.js bundle"
  }
}
```

Given the two npm packages, only one of which is _supposed_ to have [a
package bin][bin], this change makes it so there's only a single
`react-native` executable between them.

[c1]: https://github.com/facebook/react-native/commit/0a5967dbfe8629074297f8ec5d85dd5d70a91a79
[c2]: https://github.com/facebook/react-native/commit/2856be9a07582310c9e56667e8db21fb7b29ab48
[bin]: https://docs.npmjs.com/files/package.json#bin