Fix permissions on Android 6.0
The only thing blocking React Native from working on Android 6.0's
permissions model is the use of the SYSTEM_ALERT Dialog type, but
we don't necessarily need to use this. Instead, we need a
reference to the current Activity to use as the Dialog's context.
There may be a better way to do this, but simply setting a field
on the DevSupportManager is the easiest and simplest way to do it.