Adds support for JavaScriptModules and configuration writer.
One limitation of the .NET framework is that there is no equivalent of the Proxy class in Java that is capable of instantiating interfaces and specifying an invocation handler.

Instead of using interfaces in the way they were used in the Android React implementation, this implementation just uses a base class and a settable property for an invocation handler.

The configuration writer is straight-forward, it just uses the JsonWriter from Newtonsoft.Json.