Cross platform pull to refresh component.
When put inside a ScrollView it will add a UIRefreshControl on iOS and a
SwipeRefreshLayout on Android. Adds support for tintColor and title on iOS
and all the props currently supported on PullToRefreshViewAndroid on
Android.

This will allow to deprecate/remove the onRefreshStart prop on ScrollView
and PullToRefreshViewAndroid in a future PR.