Add default timeouts to android and ios
The default native stack timeouts are 60s and 120s for ios and
android respectively. This commit amends a default so that the
timeout event is triggered rather than allowing the native stack
to timeout and fail silently. It is still possible that the
timeout could be set to greater than the default resulting in it
failing silently.